requirements.txt
configparser
sendsms
django-sendsms
